+++ CHANGELOG.md
@@ -3,8 +3,12 @@
 ## Version 0.12.7
 
 - feat(notification): make the extra information screen scrollable
-- build(bootstrap): add `dtoverlay=gpio-fan,gpiopin=22,temp=60000` to `/boot/config.txt`
+- build(bootstrap): add `dtoverlay=gpio-fan,gpiopin=22,temp=60000` to `/boot/firmware/config.txt`
   to make the fan run if CPU temperature passes 60℃ #64
+- fix(audio): run the original `install.sh` script of `wm8960-soundcard` to make
+  the audio work #53
+- build(packer): mount the first partition of the image in `/boot/firmware` instead
+  of `/boot` to be compatible with the new linux kernel
 
 ## Version 0.12.6